+++ GameSpeed2U/src/order/history/OrderHistory.java
@@ -1,4 +1,6 @@
-package order;
+package order.history;
+
+import order.order.OrderItem;
 
 import java.util.ArrayList;
