+++ GameSpeed2U/src/order/order/OrdersHashMap.java
@@ -1,4 +1,4 @@
-package order;
+package order.order;
 
 import java.util.ArrayList;
 import java.util.HashMap;
