+++ GameSpeed2U/src/user/account/Logout.java
@@ -1,4 +1,4 @@
-package user;
+package user.account;
 
 import main.Helper;
