+++ GameSpeed2U/src/user/info/UserPersonalInfoPage.java
@@ -1,4 +1,4 @@
-package user;
+package user.info;
 
 import main.Helper;
