+++ GameSpeed2U/src/user/review/Review.java
@@ -1,4 +1,4 @@
-package user;
+package user.review;
 
 import main.Helper;
