+++ GameSpeed2U/src/user/review/ReviewForm.java
@@ -1,4 +1,4 @@
-package user;
+package user.review;
 
 /**
  1. ProductModelName: XBOX 360
